feat: extract components from NFTBrowse
