--- webapp/src/components/CategoriesMenu/index.ts
@@ -1,2 +1,0 @@
-import CategoriesMenu from './CategoriesMenu.container'
-export { CategoriesMenu }
