--- webapp/src/components/NFTBrowse/FiltersMenu/FiltersMenu.css
@@ -1,2 +1,0 @@
-.FiltersMenu {
-}
