+++ webapp/src/components/NFTFilters/NFTFilters.css
@@ -1,0 +1,65 @@
+.FiltersModal .CategoriesSidebar ul.menu li:hover,
+.FiltersModal .CategoriesSidebar ul.menu li.active {
+  background-color: var(--secondary-on-modal);
+}
+
+.FiltersModal .filter-row {
+  margin-bottom: 20px;
+}
+
+.FiltersModal .filter-row .ui.sub.header {
+  margin-bottom: 12px;
+}
+
+.FiltersModal .CategoriesSidebar {
+  width: 100%;
+}
+
+.FiltersModal .CategoriesSidebar ul.menu {
+  width: calc(100% + 16px);
+}
+
+.FiltersModal .CategoriesSidebar ul.menu li {
+  width: calc(100% + 16px);
+}
+
+.FiltersModal .CategoriesSidebar ul.menu li.sub {
+  width: calc(100% + 0px);
+}
+
+.FiltersModal.ui.modal .dcl.close {
+  position: absolute;
+  right: 24px !important;
+  top: 24px !important;
+}
+
+.FiltersModal .apply-filters {
+  width: 100%;
+  margin-top: 48px;
+}
+
+.FiltersModal .ui.dropdown {
+  display: flex;
+  align-items: center;
+  padding: 6px 10px;
+  border: 1px solid var(--divider);
+  border-radius: 6px;
+  color: var(--secondary-text);
+}
+
+.FiltersModal .ui.dropdown .text {
+  flex: 1 0 auto;
+  color: var(--text);
+}
+
+.FiltersModal .ui.dropdown .dropdown.icon {
+  color: var(--text);
+}
+
+@media (max-width: 768px) {
+  .FiltersModal .dcl.row {
+    align-items: flex-start;
+    flex-direction: column;
+    justify-content: stretch;
+  }
+}
