+++ webapp/src/components/NFTFilters/index.ts
@@ -1,0 +1,2 @@
+import NFTFilters from './NFTFilters.container'
+export { NFTFilters }
