+++ webapp/src/components/NFTList/NFTList.types.ts
@@ -1,0 +1,10 @@
+import { NFT } from '../../modules/nft/types'
+
+export type Props = {
+  nfts: NFT[]
+  page: number
+  count?: number
+  isLoading: boolean
+}
+
+export type MapStateProps = Props
