+++ webapp/src/components/NFTList/index.ts
@@ -1,0 +1,2 @@
+import NFTList from './NFTList.container'
+export { NFTList }
