+++ webapp/src/components/NFTSidebar/CategoriesSidebar/index.ts
@@ -1,0 +1,2 @@
+import CategoriesSidebar from './CategoriesSidebar.container'
+export { CategoriesSidebar }
