+++ webapp/src/components/NFTSidebar/NFTSidebar.types.ts
@@ -1,0 +1,7 @@
+import { Section } from '../../modules/routing/search'
+
+export type Props = {
+  section: Section
+}
+
+export type MapStateProps = Props
