+++ webapp/src/components/NFTSidebar/index.ts
@@ -1,0 +1,2 @@
+import NFTSidebar from './NFTSidebar.container'
+export { NFTSidebar }
