--- webapp/src/components/PartnerPage/PartnerPage.css
@@ -1,207 +1,0 @@
-.NFTListPage .right .ui.toggle.checkbox input.hidden + label {
-  font-size: 14px;
-  margin-right: 0px;
-}
-
-.NFTListPage .ui.toggle.checkbox label,
-.NFTListPage .ui.dropdown > .text {
-  text-transform: uppercase;
-}
-
-.NFTListPage .ui.cards {
-  position: relative;
-  overflow: hidden;
-}
-
-.NFTListPage .overlay {
-  height: 100%;
-  width: 100%;
-  position: absolute;
-  background: var(--background);
-  opacity: 0.6;
-}
-
-.NFTListPage .load-more {
-  text-align: center;
-  margin-top: 36px;
-  height: 46px;
-}
-
-.NFTListPage .empty {
-  margin-top: 40px;
-  font-size: 16px;
-  text-align: center;
-  color: var(--secondary-text);
-}
-
-.NFTListPage .results {
-  margin: 0px;
-}
-
-.NFTListPage .filters {
-  padding: 16px;
-  background-color: var(--card);
-  border-radius: 16px;
-  margin-bottom: 32px;
-  user-select: none;
-  outline: none;
-  padding-bottom: 8px;
-  margin-top: -8px;
-}
-
-.NFTListPage .filters .Filter + .Filter {
-  margin-top: 14px;
-}
-
-.NFTListPage .filters .dcl.row {
-  align-items: flex-start;
-}
-
-.NFTListPage .filters .dcl.row + .dcl.row {
-  margin-top: 12px;
-}
-
-.NFTListPage .filters .dcl.row .Filter {
-  flex: 1 0 auto;
-}
-
-.NFTListPage .filters .dcl.row .Filter + .Filter {
-  margin-top: 0px;
-  flex: none;
-  width: 50%;
-  margin-left: 24px;
-}
-
-.NFTListPage .open-filters-wrapper {
-  display: inline-block;
-  height: 15px;
-  margin-left: 24px;
-  position: relative;
-  cursor: pointer;
-}
-
-.NFTListPage .open-filters-wrapper .label {
-  color: var(--primary);
-  text-transform: uppercase;
-  margin-right: 30px;
-  font-weight: 500;
-}
-
-.NFTListPage .open-filters {
-  width: 20px;
-  height: 20px;
-  background-image: url(../../images/filters.svg);
-  background-size: 16px;
-  background-repeat: no-repeat;
-  background-position: center;
-
-  display: inline-block;
-  position: absolute;
-  right: 0px;
-  top: 0px;
-}
-
-.NFTListPage .open-filters.active {
-  background-image: url(../../images/filters-active.svg);
-  background-size: 20px;
-}
-
-.NFTListPage .topbar {
-  display: flex;
-  align-items: center;
-  margin-bottom: 4px;
-}
-
-.NFTListPage .topbar > .TextFilter {
-  flex: 1 0 auto;
-  margin-right: 32px;
-}
-
-.NFTListPage .topbar > .ui.dropdown,
-.NFTListPage .topbar > .open-filters-wrapper {
-  height: 29px;
-}
-
-.FiltersModal .CategoriesMenu ul.menu li:hover,
-.FiltersModal .CategoriesMenu ul.menu li.active {
-  background-color: var(--secondary-on-modal);
-}
-
-.FiltersModal .filter-row {
-  margin-bottom: 20px;
-}
-
-.FiltersModal .filter-row .ui.sub.header {
-  margin-bottom: 12px;
-}
-
-.FiltersModal .CategoriesMenu {
-  width: 100%;
-}
-
-.FiltersModal .CategoriesMenu ul.menu {
-  width: calc(100% + 16px);
-}
-
-.FiltersModal .CategoriesMenu ul.menu li {
-  width: calc(100% + 16px);
-}
-
-.FiltersModal .CategoriesMenu ul.menu li.sub {
-  width: calc(100% + 0px);
-}
-
-.FiltersModal.ui.modal .dcl.close {
-  position: absolute;
-  right: 24px !important;
-  top: 24px !important;
-}
-
-.FiltersModal .apply-filters {
-  width: 100%;
-  margin-top: 48px;
-}
-
-.FiltersModal .ui.dropdown {
-  display: flex;
-  align-items: center;
-  padding: 6px 10px;
-  border: 1px solid var(--divider);
-  border-radius: 6px;
-  color: var(--secondary-text);
-}
-
-.FiltersModal .ui.dropdown .text {
-  flex: 1 0 auto;
-  color: var(--text);
-}
-
-.FiltersModal .ui.dropdown .dropdown.icon {
-  color: var(--text);
-}
-
-.NFTListPage .Filter {
-  display: flex;
-  flex-direction: column;
-}
-
-.NFTListPage .Filter .ui.sub.header.name {
-  margin-bottom: 12px !important;
-  font-weight: 600;
-}
-
-@media (max-width: 768px) {
-  .NFTListPage .Layout .right.column > div {
-    width: 100%;
-  }
-
-  .CategoriesMenu ul.menu li {
-    width: calc(100% - 32px);
-  }
-
-  .FiltersModal .dcl.row {
-    align-items: flex-start;
-    flex-direction: column;
-    justify-content: stretch;
-  }
-}
