--- webapp/src/components/PartnersMenu/index.ts
@@ -1,2 +1,0 @@
-import PartnersMenu from './PartnersMenu'
-export { PartnersMenu }
