+++ webapp/src/components/PartnersSidebar/index.ts
@@ -1,0 +1,2 @@
+import PartnersSidebar from './PartnersSidebar'
+export { PartnersSidebar }
